Bugfix/wordcounts page loads slow (#67)
* preprocess data and render results only on wordcounts page

* extract page-creators from gatsby-node.js

* resolve path issues
