+++ site/gatsby-site/constants/customStopWords.js
@@ -1,0 +1,57 @@
+const stopword = require('stopword');
+
+const englishStopwords = stopword.en;
+
+const customStopWords = [
+  'not',
+  'when',
+  'time',
+  'will',
+  'report',
+  'work',
+  'sai',
+  'just',
+  'year',
+  'first',
+  'result',
+  'look',
+  'name',
+  'even',
+  'mai',
+  'problem',
+  'she',
+  'dai',
+  'case',
+  'show',
+  'two',
+  'order',
+  'includ',
+  'issu',
+  'accord',
+  'down',
+  'last',
+  'that',
+  'off',
+  'like',
+  'week',
+  'caus',
+  'suggest',
+  'happen',
+  'respons',
+  'ask',
+  'word',
+  'need',
+  'man',
+  'back',
+  'thing',
+  'differ',
+  'said',
+  'inform',
+  'servic',
+  'stop',
+  'set',
+  'question',
+  'help',
+].concat(englishStopwords);
+
+module.exports = customStopWords;
