+++ site/gatsby-site/page-creators/createMdxPages.js
@@ -1,0 +1,45 @@
+const path = require('path');
+
+const createMdxPages = (graphql, createPage) => {
+  return new Promise((resolve, reject) => {
+    resolve(
+      graphql(
+        `
+          {
+            allMdx {
+              edges {
+                node {
+                  fields {
+                    id
+                  }
+                  tableOfContents
+                  fields {
+                    slug
+                  }
+                }
+              }
+            }
+          }
+        `
+      ).then(result => {
+        if (result.errors) {
+          console.log(result.errors); // eslint-disable-line no-console
+          reject(result.errors);
+        }
+
+        // Create blog posts pages.
+        result.data.allMdx.edges.forEach(({ node }) => {
+          createPage({
+            path: node.fields.slug ? node.fields.slug : '/',
+            component: path.resolve('./src/templates/docs.js'),
+            context: {
+              id: node.fields.id,
+            },
+          });
+        });
+      })
+    );
+  });
+}
+
+module.exports = createMdxPages;
